Refactor modules: extract `partial_array::iter`
